no timeout for db migrations
