+++ packages/shared/lib/db/database.ts
@@ -1,12 +1,30 @@
 import knex from 'knex';
 import type { Knex } from 'knex';
-import { config } from './config.js';
 
-class KnexDatabase {
+function getDbConfig({ timeoutMs }: { timeoutMs: number }): Knex.Config<any> {
+    return {
+        client: process.env['NANGO_DB_CLIENT'] || 'pg',
+        connection: process.env['NANGO_DATABASE_URL'] || {
+            host: process.env['NANGO_DB_HOST'] || (process.env['SERVER_RUN_MODE'] === 'DOCKERIZED' ? 'nango-db' : 'localhost'),
+            port: +(process.env['NANGO_DB_PORT'] || 5432),
+            user: process.env['NANGO_DB_USER'] || 'nango',
+            database: process.env['NANGO_DB_NAME'] || 'nango',
+            password: process.env['NANGO_DB_PASSWORD'] || 'nango',
+            ssl: process.env['NANGO_DB_SSL'] != null && process.env['NANGO_DB_SSL'].toLowerCase() === 'true' ? { rejectUnauthorized: false } : undefined,
+            statement_timeout: timeoutMs
+        },
+        pool: {
+            min: parseInt(process.env['NANGO_DB_POOL_MIN'] || '2'),
+            max: parseInt(process.env['NANGO_DB_POOL_MAX'] || '7')
+        }
+    };
+}
+
+export class KnexDatabase {
     knex: Knex;
 
-    constructor() {
-        const dbConfig = config.development;
+    constructor({ timeoutMs } = { timeoutMs: 60000 }) {
+        const dbConfig = getDbConfig({ timeoutMs });
         this.knex = knex(dbConfig);
     }
